@@ -1,50 +1,61 @@
 package net.jmb19905.medievalworlds.client.screen;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.PoseStack;
 import net.jmb19905.medievalworlds.MedievalWorlds;
-import net.jmb19905.medievalworlds.client.containers.BloomeryContainer;
-import net.minecraft.client.gui.screen.inventory.ContainerScreen;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.ITextComponent;
+import net.jmb19905.medievalworlds.client.menus.BloomeryMenu;
+import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Inventory;
 
-public class BloomeryScreen extends ContainerScreen<BloomeryContainer> {
+import javax.annotation.Nonnull;
+
+public class BloomeryScreen extends AbstractContainerScreen<BloomeryMenu> {
 
     private static final ResourceLocation TEXTURE = new ResourceLocation(MedievalWorlds.MOD_ID, "textures/gui/bloomery.png");
-    private final BloomeryContainer screenContainer;
+    private final BloomeryMenu screenContainer;
 
-    public BloomeryScreen(BloomeryContainer screenContainer, PlayerInventory inv, ITextComponent titleIn) {
+    private final int left;
+    private final int top;
+    private final int width_;
+    private final int height_;
+
+    public BloomeryScreen(BloomeryMenu screenContainer, Inventory inv, Component titleIn) {
         super(screenContainer, inv, titleIn);
         this.screenContainer = screenContainer;
 
-        this.guiLeft = 0;
-        this.guiTop = 0;
-        this.xSize = 176;
-        this.ySize = 166;
+        this.leftPos = 0;
+        this.topPos = 0;
+        this.width = 176;
+        this.height = 166;
+
+        this.left = leftPos;
+        this.top = topPos;
+        this.width_ = width;
+        this.height_ = height;
     }
 
     @Override
-    protected void drawGuiContainerBackgroundLayer(MatrixStack matrixStack, float partialTicks, int mouseX, int mouseY) {
-        RenderSystem.color4f(1, 1, 1, 1);
-        assert this.minecraft != null;
-        this.minecraft.getTextureManager().bindTexture(TEXTURE);
-        this.blit(matrixStack, this.guiLeft, this.guiTop, 0, 0, this.xSize, this.ySize);
-        this.blit(matrixStack, this.guiLeft + 79, this.guiTop + 35, 176, 14, screenContainer.getSmeltProgressionScaled(), 16);
-        this.blit(matrixStack, this.guiLeft + 39, this.guiTop + 35 + screenContainer.getBurnProgressionScaled(), 176, screenContainer.getBurnProgressionScaled(), 14, screenContainer.getCurrentMaxBurnTime() != 0 ? 14 - screenContainer.getBurnProgressionScaled() : 0);
+    protected void renderBg(@Nonnull PoseStack stack, float partialTicks, int mouseX, int mouseY) {
+        RenderSystem.setShaderColor(1, 1, 1, 1);
+        RenderSystem.setShaderTexture(0, TEXTURE);
+        this.blit(stack, this.left, this.top, 0, 0, this.width_, this.height_);
+        this.blit(stack, this.left + 79, this.top + 35, 176, 14, screenContainer.getSmeltProgressionScaled(), 16);
+        this.blit(stack, this.left + 39, this.top + 35 + screenContainer.getBurnProgressionScaled(), 176, screenContainer.getBurnProgressionScaled(), 14, screenContainer.getCurrentMaxBurnTime() != 0 ? 14 - screenContainer.getBurnProgressionScaled() : 0);
     }
 
     @Override
-    protected void drawGuiContainerForegroundLayer(MatrixStack matrixStack, int mouseX, int mouseY) {
-        super.drawGuiContainerForegroundLayer(matrixStack, mouseX, mouseY);
-        this.font.drawString(matrixStack, this.title.getString(), 8,5, 0x404040);
-        this.font.drawString(matrixStack, this.playerInventory.getDisplayName().getString(), 8, 72, 0x404040);
+    protected void renderLabels(@Nonnull PoseStack stack, int mouseX, int mouseY) {
+        super.renderLabels(stack, mouseX, mouseY);
+        this.font.draw(stack, this.title.getString(), 8,5, 0x404040);
+        this.font.draw(stack, this.playerInventoryTitle.getString(), 8, 72, 0x404040);
     }
 
     @Override
-    public void render(MatrixStack matrixStack, int p_render_1_, int p_render_2_, float p_render_3_) {
-        this.renderBackground(matrixStack);
-        super.render(matrixStack, p_render_1_, p_render_2_, p_render_3_);
+    public void render(@Nonnull PoseStack stack, int p_render_1_, int p_render_2_, float p_render_3_) {
+        this.renderBackground(stack);
+        super.render(stack, p_render_1_, p_render_2_, p_render_3_);
         //TODO: this.renderToolTip(p_render_1_, p_render_2_);
     }
 
