@@ -1,84 +1,82 @@
 package net.jmb19905.medievalworlds.common.capability;
 
 import net.jmb19905.medievalworlds.MedievalWorlds;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.nbt.INBT;
-import net.minecraft.util.Direction;
-import net.minecraft.util.math.vector.Vector3d;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.CapabilityInject;
+import net.minecraftforge.common.capabilities.CapabilityManager;
 
-import javax.annotation.Nullable;
+public class Motion implements IMotion{
 
-public class Motion {
+    public static final ResourceLocation ID = new ResourceLocation(MedievalWorlds.MOD_ID, "motion");
 
-    private Vector3d prevPos;
-    private Vector3d pos;
-
-    public Motion(){
-        this(Vector3d.ZERO, Vector3d.ZERO);
+    @CapabilityInject(IMotion.class)
+    public static final Capability<IMotion> MOTION_CAPABILITY = null;
+    public static void register() {
+        CapabilityManager.INSTANCE.register(IMotion.class);
     }
 
-    public Motion(Vector3d prevPos, Vector3d pos){
+    private Vec3 prevPos;
+    private Vec3 pos;
+
+    public Motion(){
+        this(Vec3.ZERO, Vec3.ZERO);
+    }
+
+    public Motion(Vec3 prevPos, Vec3 pos){
         this.prevPos = prevPos;
         this.pos = pos;
     }
 
-    public Vector3d getPrevPos() {
+    public Vec3 getPrevPos() {
         return prevPos;
     }
 
-    public Vector3d getPos() {
+    public Vec3 getPos() {
         return pos;
     }
 
-    public void setPrevPos(Vector3d prevPos) {
+    public void setPrevPos(Vec3 prevPos) {
         this.prevPos = prevPos;
     }
 
-    public void setPos(Vector3d pos) {
+    public void setPos(Vec3 pos) {
         this.pos = pos;
     }
 
-    public double getMotionDistance(){
+    @Override
+    public Vec3 getMoveVec(){
+        return pos.subtract(prevPos);
+    }
+
+    @Override
+    public double getMoveDist() {
         return pos.distanceTo(prevPos);
     }
 
-    public static class MotionNBTStorage implements Capability.IStorage<Motion> {
-        @Nullable
-        @Override
-        public INBT writeNBT(Capability<Motion> capability, Motion instance, Direction side) {
-            CompoundNBT nbt = new CompoundNBT();
-            nbt.putDouble(MedievalWorlds.MOD_ID + "_prevPos_x", instance.prevPos.x);
-            nbt.putDouble(MedievalWorlds.MOD_ID + "_prevPos_y", instance.prevPos.y);
-            nbt.putDouble(MedievalWorlds.MOD_ID + "_prevPos_x", instance.prevPos.x);
-            nbt.putDouble(MedievalWorlds.MOD_ID + "_pos_x", instance.pos.x);
-            nbt.putDouble(MedievalWorlds.MOD_ID + "_pos_y", instance.pos.y);
-            nbt.putDouble(MedievalWorlds.MOD_ID + "_pos_x", instance.pos.x);
-            return nbt;
-        }
-
-        @Override
-        public void readNBT(Capability<Motion> capability, Motion instance, Direction side, INBT nbt) {
-            double prevX = 0;
-            double prevY = 0;
-            double prevZ = 0;
-            double x = 0;
-            double y = 0;
-            double z = 0;
-            if(nbt.getType() == CompoundNBT.TYPE){
-                prevX = ((CompoundNBT) nbt).getDouble(MedievalWorlds.MOD_ID + "_prevPos_x");
-                prevY = ((CompoundNBT) nbt).getDouble(MedievalWorlds.MOD_ID + "_prevPos_y");
-                prevZ = ((CompoundNBT) nbt).getDouble(MedievalWorlds.MOD_ID + "_prevPos_z");
-                x = ((CompoundNBT) nbt).getDouble(MedievalWorlds.MOD_ID + "_pos_x");
-                y = ((CompoundNBT) nbt).getDouble(MedievalWorlds.MOD_ID + "_pos_y");
-                z = ((CompoundNBT) nbt).getDouble(MedievalWorlds.MOD_ID + "_pos_z");
-            }
-            instance.setPrevPos(new Vector3d(prevX,prevY,prevZ));
-            instance.setPos(new Vector3d(x,y,z));
-        }
+    @Override
+    public CompoundTag serializeNBT() {
+        CompoundTag nbt = new CompoundTag();
+        nbt.putDouble(MedievalWorlds.MOD_ID + ".motionPrevX", getPrevPos().x());
+        nbt.putDouble(MedievalWorlds.MOD_ID + ".motionPrevY", getPrevPos().x());
+        nbt.putDouble(MedievalWorlds.MOD_ID + ".motionPrevZ", getPrevPos().x());
+        nbt.putDouble(MedievalWorlds.MOD_ID + ".motionX", getPos().x());
+        nbt.putDouble(MedievalWorlds.MOD_ID + ".motionX", getPos().x());
+        nbt.putDouble(MedievalWorlds.MOD_ID + ".motionX", getPos().x());
+        return nbt;
     }
 
-    public static Motion createDefaultInstance() {
-        return new Motion();
+    @Override
+    public void deserializeNBT(CompoundTag nbt) {
+        double prevX = nbt.getDouble(MedievalWorlds.MOD_ID + "motionPrevX");
+        double prevY = nbt.getDouble(MedievalWorlds.MOD_ID + "motionPrevY");
+        double prevZ = nbt.getDouble(MedievalWorlds.MOD_ID + "motionPrevZ");
+        setPrevPos(new Vec3(prevX, prevY, prevZ));
+        double x = nbt.getDouble(MedievalWorlds.MOD_ID + "motionX");
+        double y = nbt.getDouble(MedievalWorlds.MOD_ID + "motionY");
+        double z = nbt.getDouble(MedievalWorlds.MOD_ID + "motionZ");
+        setPos(new Vec3(x, y, z));
     }
 }
