@@ -2,35 +2,34 @@ package net.jmb19905.medievalworlds.common.item;
 
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Multimap;
-import net.jmb19905.medievalworlds.common.capability.MotionCapability;
+import net.jmb19905.medievalworlds.common.capability.Motion;
 import net.jmb19905.medievalworlds.common.item.lance.EntityMessageToServer;
 import net.jmb19905.medievalworlds.common.networking.NetworkStartupCommon;
 import net.jmb19905.medievalworlds.util.ConfigHandler;
 import net.jmb19905.medievalworlds.util.Util;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.entity.EntityRenderer;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.ai.attributes.Attribute;
-import net.minecraft.entity.ai.attributes.AttributeModifier;
-import net.minecraft.entity.passive.horse.AbstractHorseEntity;
-import net.minecraft.entity.passive.horse.LlamaEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.inventory.EquipmentSlotType;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.UseAction;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.math.EntityRayTraceResult;
-import net.minecraft.util.math.RayTraceResult;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.World;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.Attribute;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.animal.horse.AbstractHorse;
+import net.minecraft.world.entity.animal.horse.Llama;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.UseAnim;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.EntityHitResult;
+import net.minecraft.world.phys.HitResult;
+import net.minecraft.world.phys.Vec3;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -56,52 +55,49 @@ public class LanceItem extends Item {
     }
 
     @Override
-    public boolean onLeftClickEntity(ItemStack stack, PlayerEntity player, Entity entity) {
-        Util.attackWithStack(player, entity, 1);
-        return true;
-    }
-
-    @Override
     public int getUseDuration(@Nonnull ItemStack stack) {
         return 72000;
     }
 
-    @Override
     @Nonnull
-    public UseAction getUseAction(@Nonnull ItemStack stack) {
-        return UseAction.NONE;
+    @Override
+    public UseAnim getUseAnimation(@Nonnull ItemStack stack) {
+        return UseAnim.NONE;
     }
 
-    @Override
     @Nonnull
-    public ActionResult<ItemStack> onItemRightClick(@Nonnull World worldIn, PlayerEntity playerIn,@Nonnull  Hand hand) {
-        ItemStack itemStack = playerIn.getHeldItem(hand);
-        if(hand == Hand.MAIN_HAND) {
-            playerIn.setActiveHand(hand);
-            return new ActionResult<>(ActionResultType.SUCCESS, itemStack);
+    @Override
+    public InteractionResultHolder<ItemStack> use(@Nonnull Level level, Player player, @Nonnull InteractionHand hand) {
+        ItemStack itemStack = player.getItemInHand(hand);
+        if(hand == InteractionHand.MAIN_HAND) {
+            player.swing(hand);
+
+            return new InteractionResultHolder<>(InteractionResult.SUCCESS, itemStack);
         } else {
-            return new ActionResult<>(ActionResultType.FAIL, itemStack);
+            return new InteractionResultHolder<>(InteractionResult.FAIL, itemStack);
         }
     }
 
+    @SuppressWarnings("ConstantConditions")
     @Override
-    public void onUsingTick(ItemStack stack, LivingEntity player, int count) {
-        if(player.getRidingEntity() instanceof AbstractHorseEntity && !(player.getRidingEntity() instanceof LlamaEntity)) {
-            if (player.world.isRemote) {
-                PlayerEntity playerEntity = (PlayerEntity) player;
-                if(!playerEntity.getCooldownTracker().hasCooldown(this)) {
-                    RayTraceResult rayTrace = Minecraft.getInstance().objectMouseOver;
-                    if (rayTrace != null) {
-                        if (rayTrace.getType().equals(RayTraceResult.Type.ENTITY)) {
-                            EntityRayTraceResult entityRayTrace = (EntityRayTraceResult) rayTrace;
-                            sendDataToServer(entityRayTrace.getEntity().getEntityId(), count >= fullCharge);
+    public void onUseTick(@Nonnull Level level, LivingEntity living, @Nonnull ItemStack stack, int useTime) {
+        if(living.getRootVehicle() instanceof AbstractHorse && !(living.getRootVehicle() instanceof Llama)) {
+            if(living.level.isClientSide) {
+                Player player = (Player) living;
+                if(!player.getCooldowns().isOnCooldown(this)) {
+                    HitResult hitResult = Minecraft.getInstance().hitResult;
+                    if(hitResult != null) {
+                        if(hitResult.getType().equals(HitResult.Type.ENTITY)) {
+                            EntityHitResult entityHitResult = (EntityHitResult) hitResult;
+                            sendDataToServer(entityHitResult.getEntity().getId(), useTime >= fullCharge);
                         }
                     }
                 }
-            }else {
-                player.getCapability(MotionCapability.CAPABILITY_MOTION).ifPresent(motion -> {
+            } else {
+                living.getCapability(Motion.MOTION_CAPABILITY).ifPresent(iMotion -> {
+                    Motion motion = (Motion) iMotion;
                     motion.setPrevPos(motion.getPos());
-                    motion.setPos(player.getPositionVec());
+                    motion.setPos(new Vec3(living.getX(), living.getY(), living.getZ()));
                 });
             }
         }
@@ -113,13 +109,13 @@ public class LanceItem extends Item {
     }
 
     @Override
-    public void addInformation(ItemStack stack, @Nullable World world, List<ITextComponent> textComponents, ITooltipFlag tooltipFlag) {
-        textComponents.add(new TranslationTextComponent("tooltip.usedOnHorse"));
+    public void appendHoverText(@Nonnull ItemStack stack, @Nullable Level level, List<Component> components, @Nonnull TooltipFlag tooltipFlag) {
+        components.add(new TranslatableComponent("tooltip.usedOnHorse"));
         double minimum = roundToHalf(ConfigHandler.COMMON.lanceBaseAttackDamage.get() * 0.237 * attackDamageFactor);
-        textComponents.add(new TranslationTextComponent("tooltip.minimumAttack").append(new StringTextComponent(": " + minimum)));
+        components.add(new TranslatableComponent("tooltip.minimumAttack").append(new TextComponent(": " + minimum)));
         double maximum = roundToHalf(ConfigHandler.COMMON.lanceBaseAttackDamage.get() * 0.7115 * attackDamageFactor);
-        textComponents.add(new TranslationTextComponent("tooltip.maximumAttack").append(new StringTextComponent(": " + maximum)));
-        textComponents.add(new TranslationTextComponent("tooltip.chargeTime").append(new StringTextComponent(": " + (fullCharge / 2))));
+        components.add(new TranslatableComponent("tooltip.maximumAttack").append(new TextComponent(": " + maximum)));
+        components.add(new TranslatableComponent("tooltip.chargeTime").append(new TextComponent(": " + (fullCharge / 2))));
     }
 
     private double roundToHalf(double d) {
@@ -127,12 +123,8 @@ public class LanceItem extends Item {
     }
 
     @Override
-    public Multimap<Attribute, AttributeModifier> getAttributeModifiers(EquipmentSlotType equipmentSlot, ItemStack stack) {
+    public Multimap<Attribute, AttributeModifier> getAttributeModifiers(EquipmentSlot equipmentSlot, ItemStack stack) {
         ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
-
-        /*if (equipmentSlot == EquipmentSlotType.MAINHAND) {
-            builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Weapon modifier", this.attackDamageFactor, AttributeModifier.Operation.ADDITION));
-        }*/
         return builder.build();
     }
 
