@@ -1,10 +1,10 @@
 package net.jmb19905.medievalworlds.common.item.armor;
 
-import net.minecraft.entity.Entity;
-import net.minecraft.inventory.EquipmentSlotType;
-import net.minecraft.item.ArmorItem;
-import net.minecraft.item.IArmorMaterial;
-import net.minecraft.item.ItemStack;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.item.ArmorMaterial;
+import net.minecraft.world.item.ItemStack;
 
 import javax.annotation.Nullable;
 
@@ -12,14 +12,14 @@ public class KnightArmorLeggingsItem extends ArmorItem {
 
     private final String material;
 
-    public KnightArmorLeggingsItem(IArmorMaterial materialIn, EquipmentSlotType slot, Properties builder, String material) {
+    public KnightArmorLeggingsItem(ArmorMaterial materialIn, EquipmentSlot slot, Properties builder, String material) {
         super(materialIn, slot, builder);
         this.material = material;
     }
 
     @Nullable
     @Override
-    public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlotType slot, String type) {
+    public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlot slot, String type) {
         return "medievalworlds:textures/models/armor/"+material+"_knight_layer_2.png";
     }
 }
