@@ -1,28 +1,28 @@
 package net.jmb19905.medievalworlds.common.recipes;
 
 import net.jmb19905.medievalworlds.MedievalWorlds;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.IRecipe;
-import net.minecraft.item.crafting.IRecipeType;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.registry.Registry;
+import net.minecraft.core.Registry;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraftforge.items.wrapper.RecipeWrapper;
 
 import javax.annotation.Nonnull;
 import java.util.Objects;
 
-public interface IAlloyRecipe extends IRecipe<RecipeWrapper> {
+public interface IAlloyRecipe extends Recipe<RecipeWrapper> {
 
     ResourceLocation RECIPE_TYPE_ID = new ResourceLocation(MedievalWorlds.MOD_ID, "alloy");
 
     @Nonnull
     @Override
-    default IRecipeType<?> getType(){
-        return Objects.requireNonNull(Registry.RECIPE_TYPE.getOrDefault(RECIPE_TYPE_ID));
+    default RecipeType<?> getType(){
+        return Objects.requireNonNull(Registry.RECIPE_TYPE.get(RECIPE_TYPE_ID));
     }
 
     @Override
-    default boolean canFit(int width, int height) {
+    default boolean canCraftInDimensions(int p_43999_, int p_44000_){
         return false;
     }
 
