@@ -1,46 +1,64 @@
 package net.jmb19905.medievalworlds.common.world.gen;
 
-import net.jmb19905.medievalworlds.common.registries.BlockRegistryHandler;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.world.biome.Biome;
-import net.minecraft.world.gen.GenerationStage;
-import net.minecraft.world.gen.feature.*;
-import net.minecraft.world.gen.feature.template.RuleTest;
-import net.minecraft.world.gen.placement.IPlacementConfig;
-import net.minecraft.world.gen.placement.Placement;
-import net.minecraft.world.gen.placement.TopSolidRangeConfig;
+import com.google.common.collect.ImmutableList;
+import net.jmb19905.medievalworlds.MedievalWorlds;
+import net.minecraft.core.Registry;
+import net.minecraft.data.BuiltinRegistries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.valueproviders.UniformInt;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.levelgen.GenerationStep;
+import net.minecraft.world.level.levelgen.VerticalAnchor;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.OreConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.ReplaceBlockConfiguration;
 import net.minecraftforge.common.world.BiomeGenerationSettingsBuilder;
 import net.minecraftforge.event.world.BiomeLoadingEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.RegistryObject;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.ForgeRegistries;
+
+import java.util.ArrayList;
+
+import static net.minecraft.data.worldgen.Features.*;
 
 public class OreGenerator {
 
-    public static DeferredRegister<Feature<?>> ORE_FEATURES = DeferredRegister.create(ForgeRegistries.FEATURES, "medievalworlds");
+    public static ImmutableList<OreConfiguration.TargetBlockState> ORE_TARGET_LIST;
+    public static final ArrayList<ConfiguredFeature<?, ?>> overworldOres = new ArrayList<>();
 
-    public static final RegistryObject<Feature<ReplaceBlockConfig>> RUBY_ORE = ORE_FEATURES.register("ruby_ore", () -> new ReplaceBlockFeature(ReplaceBlockConfig.field_236604_a_));
+    public static void registerOres() {
+        ORE_TARGET_LIST = ImmutableList.of(OreConfiguration.target(OreConfiguration.Predicates.NATURAL_STONE, Blocks.STONE.defaultBlockState()));
+        overworldOres.add(register("tin_ore", Feature.ORE.configured(new OreConfiguration(ORE_COPPER_TARGET_LIST, 7))
+                .rangeTriangle(VerticalAnchor.absolute(0), VerticalAnchor.absolute(100))
+                .squared()
+                .count(3)
+        ));
+        overworldOres.add(register("silver_ore", Feature.ORE.configured(new OreConfiguration(ORE_GOLD_TARGET_LIST, 11))
+                .rangeUniform(VerticalAnchor.bottom(), VerticalAnchor.absolute(31))
+                .squared()
+                .count(3)
+        ));
+        overworldOres.add(register("ruby_ore", Feature.REPLACE_SINGLE_BLOCK.configured(new ReplaceBlockConfiguration(ORE_EMERALD_TARGET_LIST))
+                .rangeUniform(VerticalAnchor.absolute(4), VerticalAnchor.absolute(31))
+                .squared()
+                .count(UniformInt.of(3, 8))
+        ));
+    }
 
-    @SubscribeEvent
-    public static void generateOres(final BiomeLoadingEvent evt){
-        if(!evt.getCategory().equals(Biome.Category.NETHER) && !evt.getCategory().equals(Biome.Category.THEEND)) {
-            if (evt.getCategory().equals(Biome.Category.EXTREME_HILLS)) {
-                generateGem(evt.getGeneration(), BlockRegistryHandler.RUBY_ORE_BLOCK.get().getDefaultState());
+    private static <FC extends FeatureConfiguration> ConfiguredFeature<FC,?> register(String name, ConfiguredFeature<FC,?> configuredFeature) {
+        return Registry.register(BuiltinRegistries.CONFIGURED_FEATURE, new ResourceLocation(MedievalWorlds.MOD_ID, name), configuredFeature);
+    }
+
+    @SubscribeEvent(priority = EventPriority.HIGHEST)
+    public static void registerBiomeModification(final BiomeLoadingEvent evt){
+        BiomeGenerationSettingsBuilder generation = evt.getGeneration();
+        for(ConfiguredFeature<?,?> ore : overworldOres) {
+            if(ore != null) {
+                generation.getFeatures(GenerationStep.Decoration.UNDERGROUND_ORES).add(() -> ore);
             }
-            generateOre(evt.getGeneration(), OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD, BlockRegistryHandler.COPPER_ORE_BLOCK.get().getDefaultState(), 15, 0, 100,20);
-            generateOre(evt.getGeneration(), OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD, BlockRegistryHandler.TIN_ORE_BLOCK.get().getDefaultState(), 14, 0, 120,19);
-            generateOre(evt.getGeneration(), OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD, BlockRegistryHandler.SILVER_ORE_BLOCK.get().getDefaultState(), 10, 0, 40,10);
         }
-    }
-
-    private static void generateOre(BiomeGenerationSettingsBuilder generationSettings, RuleTest fillerBlockType, BlockState state, int veinSize, int minY, int maxY, int veinsPerChunk){
-        generationSettings.withFeature(GenerationStage.Decoration.UNDERGROUND_ORES, Feature.ORE.withConfiguration(new OreFeatureConfig(fillerBlockType, state, veinSize)).withPlacement(Placement.RANGE.configure(new TopSolidRangeConfig(minY, 0, maxY))).square().func_242731_b(veinsPerChunk));
-    }
-
-    private static void generateGem(BiomeGenerationSettingsBuilder generationSettings, BlockState gemBlockState){
-        generationSettings.withFeature(GenerationStage.Decoration.UNDERGROUND_ORES, RUBY_ORE.get().withConfiguration(new ReplaceBlockConfig(Blocks.STONE.getDefaultState(), gemBlockState)).withPlacement(Placement.EMERALD_ORE.configure(IPlacementConfig.NO_PLACEMENT_CONFIG)));
     }
 
 }
