@@ -1,7 +1,7 @@
 package net.jmb19905.medievalworlds.util.slots;
 
-import net.minecraft.item.ItemStack;
-import net.minecraft.world.World;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.SlotItemHandler;
 
@@ -9,15 +9,15 @@ import javax.annotation.Nonnull;
 
 public class AlloyInputSlot extends SlotItemHandler {
 
-    private final World world;
+    private final Level level;
 
-    public AlloyInputSlot(IItemHandler itemHandler, int index, int xPosition, int yPosition, World world) {
+    public AlloyInputSlot(IItemHandler itemHandler, int index, int xPosition, int yPosition, Level level) {
         super(itemHandler, index, xPosition, yPosition);
-        this.world = world;
+        this.level = level;
     }
 
     @Override
-    public boolean isItemValid(@Nonnull ItemStack stack) {
+    public boolean mayPlace(@Nonnull ItemStack stack) {
         //return AlloyFurnaceTileEntity.getAllRecipeInputs(RecipeSerializerRegistryHandler.ALLOY_TYPE, world).contains(stack);
         return true;
     }
