@@ -1,7 +1,7 @@
 package net.jmb19905.medievalworlds.util.slots;
 
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.SlotItemHandler;
 
@@ -9,7 +9,7 @@ import javax.annotation.Nonnull;
 
 public class BloomInputSlot extends SlotItemHandler {
 
-    private ItemStack currentFuelItem;
+    private final ItemStack currentFuelItem;
 
     public BloomInputSlot(IItemHandler itemHandler, int index, int fuelIndex, int xPosition, int yPosition) {
         super(itemHandler, index, xPosition, yPosition);
@@ -17,12 +17,12 @@ public class BloomInputSlot extends SlotItemHandler {
     }
 
     @Override
-    public boolean isItemValid(@Nonnull ItemStack stack) {
+    public boolean mayPlace(@Nonnull ItemStack stack) {
         return stack.getItem() == Items.IRON_ORE;
     }
 
     @Override
-    public int getSlotStackLimit() {
+    public int getMaxStackSize() {
         return currentFuelItem.getCount() * 8;
     }
 }
