@@ -1,6 +1,7 @@
 package net.jmb19905.medievalworlds.util.slots;
 
-import net.minecraft.item.ItemStack;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.SlotItemHandler;
@@ -9,13 +10,16 @@ import javax.annotation.Nonnull;
 
 public class FuelInputSlot extends SlotItemHandler {
 
-    public FuelInputSlot(IItemHandler itemHandler, int index, int xPosition, int yPosition) {
+    private final RecipeType<?> type;
+
+    public FuelInputSlot(IItemHandler itemHandler, int index, int xPosition, int yPosition, RecipeType<?> type) {
         super(itemHandler, index, xPosition, yPosition);
+        this.type = type;
     }
 
     @Override
-    public boolean isItemValid(@Nonnull ItemStack stack) {
-        return ForgeHooks.getBurnTime(stack) > 0;
+    public boolean mayPlace(@Nonnull ItemStack stack) {
+        return ForgeHooks.getBurnTime(stack, type) > 0;
     }
 
 }
